// src/types/user.types.ts
import { Model } from 'sequelize';

export interface UserRegistrationData {
  username: string;
  password: string;
  email: string;
  role: 'Student' | 'Teacher';
  first_name?: string;
  last_name?: string;
}

export interface UserCreationData {
  username: string;
  password: string;
  email: string;
  role: 'Admin' | 'Teacher' | 'Student' | 'Parent';
  school_id: string;
  first_name?: string;
  last_name?: string;
}

export interface UserAttributes {
  user_id?: string; // Optional, auto-generated by Sequelize
  school_id: string;
  role: 'Admin' | 'Teacher' | 'Student' | 'Parent';
  username: string;
  password_hash: string;
  email: string;
  first_name?: string;
  last_name?: string;
  is_approved?: boolean;
}

export interface UserInstance extends Model<UserAttributes>, UserAttributes {}